+++ pkgs/dart_mcp/lib/src/api/completions.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+/// Note that all features here are only supported since version
+/// [ProtocolVersion.v2025_03_26].
 part of 'api.dart';
 
 /// A request from the client to the server, to ask for completion options.